requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
opencv-python==4.9.0.80
numpy==1.26.4 
onnxruntime==1.19.2
Pillow==10.4.0